move to SerialPort
